+++ kabus/send_order_stock.go
@@ -20,7 +20,7 @@ type SendOrderStockRequest struct {
 	Qty                int                 `json:"Qty"`                // 注文数量
 	ClosePositionOrder ClosePositionOrder  `json:"ClosePositionOrder"` // 決済順序 ※信用取引の場合必須
 	ClosePositions     []ClosePosition     `json:"ClosePositions"`     // 返済建玉指定 ※信用取引の場合必須
-	Price              int                 `json:"Price"`              // 注文価格
+	Price              float64             `json:"Price"`              // 注文価格
 	ExpireDay          YmdNUM              `json:"ExpireDay"`          // 注文有効期限（年月日）
 	FrontOrderType     StockFrontOrderType `json:"FrontOrderType"`     // 執行条件
 }
@@ -85,7 +85,7 @@ type sendOrderStockRequestWithoutClosePositionOrder struct {
 	AccountType     AccountType         `json:"AccountType"`     // 口座種別
 	Qty             int                 `json:"Qty"`             // 注文数量
 	ClosePositions  []ClosePosition     `json:"ClosePositions"`  // 返済建玉指定
-	Price           int                 `json:"Price"`           // 注文価格
+	Price           float64             `json:"Price"`           // 注文価格
 	ExpireDay       YmdNUM              `json:"ExpireDay"`       // 注文有効期限（年月日）
 	FrontOrderType  StockFrontOrderType `json:"FrontOrderType"`  // 執行条件
 }
@@ -104,7 +104,7 @@ type sendOrderStockRequestWithoutClosePositions struct {
 	AccountType        AccountType         `json:"AccountType"`        // 口座種別
 	Qty                int                 `json:"Qty"`                // 注文数量
 	ClosePositionOrder ClosePositionOrder  `json:"ClosePositionOrder"` // 決済順序 ※信用取引の場合必須
-	Price              int                 `json:"Price"`              // 注文価格
+	Price              float64             `json:"Price"`              // 注文価格
 	ExpireDay          YmdNUM              `json:"ExpireDay"`          // 注文有効期限（年月日）
 	FrontOrderType     StockFrontOrderType `json:"FrontOrderType"`     // 執行条件
 }
